refactor(inventory): Phase 6 extract event-consumer slices
- 4 Features/<Slice>/ folders (ProductCreated, ReserveStock, CommitStock,
  ReleaseStock), each owning handler + slice DI extension; AddXxxSlice
  internally registers AddEventHandler<TEvent, THandler>.
- Program.cs per-handler AddEventHandler chain removed; slice manifest
  extended with 4 new AddXxxSlice() calls.
- IntegrationEvents/EventHandlers/ + IntegrationEvents/ folders deleted;
  ReserveStock event handler keeps StockReservedEvent construction inline
  (deliberate duplicate of ReserveByHttp per rule of three).
