+++ inventory-microservice/Inventory.Service/Features/CommitStock/CommitStockSliceExtensions.cs
@@ -1,0 +1,13 @@
+using ECommerce.Shared.Infrastructure.EventBus;
+using ECommerce.Shared.IntegrationEvents.Commands;
+
+namespace Inventory.Service.Features.CommitStock;
+
+internal static class CommitStockSliceExtensions
+{
+    public static IServiceCollection AddCommitStockSlice(this IServiceCollection services)
+    {
+        services.AddEventHandler<CommitStockCommand, CommitStockCommandHandler>();
+        return services;
+    }
+}
