+++ inventory-microservice/Inventory.Service/Features/ProductCreated/ProductCreatedHandler.cs
@@ -2,13 +2,13 @@ using ECommerce.Shared.Infrastructure.EventBus.Abstractions;
 using Inventory.Service.Contracts.Integration;
 using Inventory.Service.Domain.Abstractions;
 
-namespace Inventory.Service.IntegrationEvents.EventHandlers;
+namespace Inventory.Service.Features.ProductCreated;
 
-internal class ProductCreatedEventHandler : IEventHandler<ProductCreatedEvent>
+internal sealed class ProductCreatedHandler : IEventHandler<ProductCreatedEvent>
 {
     private readonly IInventoryStore _inventoryStore;
 
-    public ProductCreatedEventHandler(IInventoryStore inventoryStore)
+    public ProductCreatedHandler(IInventoryStore inventoryStore)
     {
         _inventoryStore = inventoryStore;
     }
