+++ inventory-microservice/Inventory.Service/Features/ReleaseStock/ReleaseStockCommandHandler.cs
@@ -6,15 +6,15 @@ using Inventory.Service.Contracts.Integration;
 using Inventory.Service.Domain;
 using Inventory.Service.Domain.Abstractions;
 
-namespace Inventory.Service.IntegrationEvents.EventHandlers;
+namespace Inventory.Service.Features.ReleaseStock;
 
-internal sealed class CommitStockCommandHandler : IEventHandler<CommitStockCommand>
+internal sealed class ReleaseStockCommandHandler : IEventHandler<ReleaseStockCommand>
 {
     private readonly IInventoryStore _inventoryStore;
     private readonly IOutboxUnitOfWork _outboxUnitOfWork;
     private readonly MetricFactory _metricFactory;
 
-    public CommitStockCommandHandler(
+    public ReleaseStockCommandHandler(
         IInventoryStore inventoryStore,
         IOutboxUnitOfWork outboxUnitOfWork,
         MetricFactory metricFactory)
@@ -24,31 +24,26 @@ internal sealed class CommitStockCommandHandler : IEventHandler<CommitStockComma
         _metricFactory = metricFactory;
     }
 
-    public async Task Handle(CommitStockCommand command)
+    public async Task Handle(ReleaseStockCommand command)
     {
         await _outboxUnitOfWork.ExecuteAsync(async () =>
         {
-            var result = await _inventoryStore.CommitReservations(command.OrderId);
+            var result = await _inventoryStore.ReleaseReservations(command.OrderId);
 
-            if (!result.Committed)
-            {
-                return [];
-            }
-
-            if (!result.AlreadyProcessed)
+            if (result.Released && !result.AlreadyProcessed)
             {
                 foreach (var _ in result.Lines)
                 {
                     _metricFactory.Counter("stock-movements", "movements")
-                        .Add(1, new KeyValuePair<string, object?>("movement_type", nameof(MovementType.Commit)));
+                        .Add(1, new KeyValuePair<string, object?>("movement_type", nameof(MovementType.Release)));
                 }
             }
 
             var published = result.Lines
-                .Select(l => new CommittedItem(l.ProductId, l.WarehouseId, l.Quantity))
+                .Select(l => new ReleasedItem(l.ProductId, l.WarehouseId, l.Quantity))
                 .ToList();
 
-            return [new StockCommittedEvent(command.OrderId, published)
+            return [new StockReleasedEvent(command.OrderId, published)
             {
                 CorrelationId = command.CorrelationId,
                 CausationId = command.Id,
