+++ inventory-microservice/Inventory.Service/Features/ReserveStock/ReserveStockSliceExtensions.cs
@@ -1,0 +1,13 @@
+using ECommerce.Shared.Infrastructure.EventBus;
+using ECommerce.Shared.IntegrationEvents.Commands;
+
+namespace Inventory.Service.Features.ReserveStock;
+
+internal static class ReserveStockSliceExtensions
+{
+    public static IServiceCollection AddReserveStockSlice(this IServiceCollection services)
+    {
+        services.AddEventHandler<ReserveStockCommand, ReserveStockCommandHandler>();
+        return services;
+    }
+}
